Transfered mailing functionality
An automated email will be sent out once a new user is created. The email is sent to an employees contact email adress.
