+++ app/Http/Controllers/EmployeeController.php
@@ -106,22 +106,28 @@ class EmployeeController extends Controller
     }
 
 
-    public function sendMail()
-    {
-        $login_id = Auth::id();
-        $user = DB::table('darbinieki')->where('user_id', $login_id)->first();
-
-        $to_name = $user->vards;
-//        $to_email = $user->epasts; //izkomentēts, lai varētu pārbaudīt, ka strādā, izmantojot nākošo līniju
-        $to_email = 'anitra.beinare@gmail.com';
-        $data = array('name'=> $to_name, 'body' => "Test mail");
-
-        Mail::send('emails/testMail', $data, function($message) use ($to_name, $to_email) {
-            $message->to($to_email, $to_name)
-                ->subject('RiMSIS pieslēguma izveidošana');
-            $message->from('rimsislv@gmail.com','RiMSIS Administrators');
-        });
-
-        return $this->showProfile();
-    }
+//    public function sendMail()
+//    {
+//        $login_id = Auth::id();
+//        $user = DB::table('darbinieki')->where('user_id', $login_id)->first();
+//
+//        if($user->otrais_vards != NULL)
+//        {
+//            $to_name = $user->vards . ' ' . $user->otrais_vards . ' ' . $user->uzvards;
+//        } else {
+//            $to_name = $user->vards . ' ' . $user->uzvards;
+//        }
+//
+////        $to_email = $user->epasts; //izkomentēts, lai varētu pārbaudīt, ka strādā, izmantojot nākošo līniju
+//        $to_email = 'anitra.beinare@gmail.com';
+//        $data = array('name'=> $to_name, 'body' => "Test mail");
+//
+//        Mail::send('emails/testMail', $data, function($message) use ($to_name, $to_email) {
+//            $message->to($to_email, $to_name)
+//                ->subject('RiMSIS pieslēguma izveidošana');
+//            $message->from('rimsislv@gmail.com','RiMSIS Administrators');
+//        });
+//
+//        return redirect()->route('user.show', ['id' => $user->id]);
+//    }
 }
