+++ resources/views/emails/testMail.blade.php
@@ -1,5 +1,6 @@
 <div>
     <h4>Labdien, {{ $name }}!</h4>
     <p>Sveicināts(-a) RiMSISā!</p>
-    <p>Lai pieslēgtos sistēmai, nospied uz <a href="galadarbs.test!">RiMSIS</a>!</p>
+    <p>Lai pieslēgtos sistēmai, nospied uz pogas!</p>
+    <p class="btn"><a href="galadarbs.test!">RiMSIS</a></p>
 </div>
